+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 3.0.1
+* Mopinion Android SDK updated to latest version [2.0.3](https://github.com/Mopinion-com/mopinion-sdk-android)
+* Safe cast of FLutterFragmentActivity issue fixed.
+
+## 3.0.0
+* Mopinion Android SDK updated to latest version [2.0.0](https://github.com/Mopinion-com/mopinion-sdk-android)
+* Using UIViewController instead of FlutterViewController for iOS.
+
+## 2.1.3
+* Mopinion Android SDK updated to latest version [1.0.27](https://github.com/Mopinion-com/mopinion-sdk-android)
+
+## 2.1.2
+* Mopinion Android SDK updated to latest version [1.0.26](https://github.com/Mopinion-com/mopinion-sdk-android)
+
 ## 2.1.1
 * Mopinion Android SDK updated to latest version [1.0.25](https://github.com/Mopinion-com/mopinion-sdk-android)